requirements.txt
python-dotenv
pandas
numpy
dnspython
pymongo
certifi
scikit-learn
scipy-stubs
dill
pyyaml

#-e .